requirements.txt
camelot-py==0.10.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.37.1
fpdf==1.7.2
ghostscript==0.7
idna==3.3
iopath==0.1.10
Jinja2==3.1.2
kiwisolver==1.4.4
layoutparser==0.3.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.2
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pdf2image==1.16.0
pdfminer.six==20220524
pdfplumber==0.7.4
Pillow==9.2.0
poppler-utils==0.1.0
portalocker==2.5.1
pycparser==2.21
pyparsing==3.0.9
PyPDF2==2.10.4
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scipy==1.9.1
six==1.16.0
tabulate==0.8.10
tika==1.24
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
Wand==0.6.10
